add: 채팅방 고정, 읽음 처리
